+++ Gemfile
@@ -46,6 +46,7 @@ gem "omniauth-rails_csrf_protection"
 gem 'activerecord-session_store'
 gem 'figaro'
 gem 'geocoder', '~> 1.6', '>= 1.6.7'
+gem 'ransack'
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
