+++ app/javascript/stylesheets/application.scss
@@ -1,3 +1,5 @@
 @import 'tailwindcss/base';
 @import 'tailwindcss/components';
 @import 'tailwindcss/utilities';
+@import 'easy-autocomplete/dist/easy-autocomplete';
+@import 'easy-autocomplete/dist/easy-autocomplete.themes';
