+++ app/views/events/_published.html.erb
@@ -1,6 +1,4 @@
-
 <% @events.each do |event| %>
-
 <div class="w-full bg-white text-center max-w-5xl m-auto">
     <div class="grid mt-8 gap-8 grid-cols-1 md:w-1/2 md:float-left">
         <a href="<%= event_path(event.id) %>">
