+++ app/views/events/search.json.jbuilder
@@ -1,0 +1,4 @@
+json.array!(@events) do |event|
+  json.title event.title
+  json.url event_path(event.id)
+end
