+++ package.json
@@ -20,6 +20,7 @@
     "axios": "^0.21.1",
     "datatables.net-dt": "^1.10.25",
     "datatables.net-responsive": "^2.2.8",
+    "easy-autocomplete": "^1.3.5",
     "image-map": "^2.0.1",
     "jquery": "^3.6.0",
     "postcss": "^7",
